% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discog.R
\docType{data}
\name{discog}
\alias{discog}
\title{Sharla Gelfand's music collection}
\format{A unnamed list with 155 components, each representing an item in
Sharla's music collection.}
\source{
Retrieved on 2019-07-15 from \url{https://www.discogs.com}.
}
\usage{
discog
}
\description{
Inspired by \url{https://sharla.party/posts/discog-purrr/}.
}
\examples{
length(discog)

str(discog, max.level = 2, list.len = 2)
}
\keyword{datasets}
