% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discog.R
\name{discog_json}
\alias{discog_json}
\title{Path to Discogs data as JSON}
\usage{
discog_json()
}
\value{
Local path to JSON file containing Discogs data
}
\description{
Path to Discogs data as JSON
}
\examples{
discog_json()
if (require("jsonlite")) {
  d <- fromJSON(discog_json(), simplifyVector = FALSE)
  identical(discog, d)
}
}
\seealso{
Other Discogs data and functions: 
\code{\link{discog}}
}
\concept{Discogs data and functions}
