% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gap_simple}
\alias{gap_simple}
\alias{gap_nested}
\alias{gap_split}
\title{Gapminder data frame in various forms}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1704 rows and 6 columns.
}
\usage{
gap_simple

gap_nested

gap_split
}
\description{
The main data frame from the \code{gapminder} package in three forms:
\enumerate{
\item \code{gap_simple}, same as \code{gapminder::gapminder}
\item \code{gap_nested}, nested by country and continent
\item \code{gap_split}, split by country
}
}
\examples{
gap_simple
gap_nested

str(gap_split, max.level = 1, list.len = 10)
str(gap_split[[1]])
}
\keyword{datasets}
