% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_users_repos.R
\docType{data}
\name{gh_repos}
\alias{gh_repos}
\title{GitHub repos}
\format{A unnamed list with 6 components, each itself a list of 30 repos for
  a specific GitHub user. Each repo's component is a list of length >60,
  containing information such as name, owner (a list), fork status, and
  creation date.}
\source{
\url{https://developer.github.com/v3/repos/#list-user-repositories}
}
\usage{
gh_repos
}
\description{
Info on GitHub repos, retrieved from the GitHub API.
}
\examples{
gh_repos
str(lapply(gh_repos[[1]][1:3], `[`, c("full_name", "created_at")))
}
\seealso{
Other GitHub data and functions: \code{\link{gh_users_json}},
  \code{\link{gh_users}}
}
\keyword{datasets}
