% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_users_repos.R
\name{gh_repos_json}
\alias{gh_repos_json}
\title{Path to GitHub repo JSON}
\usage{
gh_repos_json()
}
\value{
Local path to JSON file containing GitHub repo info
}
\description{
Path to GitHub repo JSON
}
\examples{
gh_repos_json()
if (require("jsonlite")) {
  fromJSON(gh_repos_json(), simplifyDataFrame = FALSE)
}
}
\seealso{
Other GitHub data and functions: \code{\link{gh_repos}},
  \code{\link{gh_users_json}}, \code{\link{gh_users}}
}

