% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_users_repos.R
\docType{data}
\name{gh_users}
\alias{gh_users}
\title{GitHub users}
\format{
A unnamed list with 6 components, each representing a GitHub user.
Each user's component is a list of length 30, containing information such
as username, GitHub id, and join date.
}
\source{
\url{https://developer.github.com/v3/users/#get-a-single-user}
}
\usage{
gh_users
}
\description{
Info on GitHub users, retrieved from the GitHub API.
}
\examples{
str(gh_users, max.level = 1)
str(gh_users[[1]])

str(lapply(gh_users, `[`, c("login", "name")))
}
\seealso{
Other GitHub data and functions: 
\code{\link{gh_repos}},
\code{\link{gh_users_json}()}
}
\concept{GitHub data and functions}
\keyword{datasets}
