% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_users_repos.R
\name{gh_users_json}
\alias{gh_users_json}
\title{Path to GitHub user JSON}
\usage{
gh_users_json()
}
\value{
Local path to JSON file containing GitHub user info
}
\description{
Path to GitHub user JSON
}
\examples{
gh_users_json()
if (require("jsonlite")) {
  fromJSON(gh_users_json(), simplifyDataFrame = FALSE)
}
}
\seealso{
Other GitHub data and functions: \code{\link{gh_repos_json}},
  \code{\link{gh_repos}}, \code{\link{gh_users}}
}

