% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_chars.R
\docType{data}
\name{got_chars}
\alias{got_chars}
\title{Game of Thrones POV characters}
\format{A unnamed list with 30 components, each representing a POV character.
Each character's component is a named list of length 18, containing
information such as name, aliases, and house allegiances.}
\source{
\url{https://anapioficeandfire.com}
}
\usage{
got_chars
}
\description{
Info on the point-of-view (POV) characters from the first five books in the
Song of Ice and Fire series by George R. R. Martin. Retrieved from An API Of
Ice And Fire.
}
\examples{
got_chars
str(lapply(got_chars, `[`, c("name", "culture")))
}
\seealso{
Other Game of Thrones data and functions: \code{\link{got_chars_json}}
}
\concept{Game of Thrones data and functions}
\keyword{datasets}
