% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/got_chars.R
\name{got_chars_json}
\alias{got_chars_json}
\alias{got_chars_xml}
\title{Paths to Game of Thrones data as JSON and XML}
\usage{
got_chars_json()

got_chars_xml()
}
\value{
Local path to JSON or XML file containing Game of Thrones data
}
\description{
Paths to Game of Thrones data as JSON and XML
}
\examples{
got_chars_json()
if (require("jsonlite")) {
  gotcj <- fromJSON(got_chars_json(), simplifyDataFrame = FALSE)
  identical(got_chars, gotcj)
}
got_chars_xml()
if (require("xml2")) {
  xml <- read_xml(got_chars_xml())
  xml
}
}
\seealso{
Other Game of Thrones data and functions: \code{\link{got_chars}}
}

