% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repurrrsive-package.r
\docType{package}
\name{repurrrsive-package}
\alias{repurrrsive}
\alias{repurrrsive-package}
\title{repurrrsive: Examples of Recursive Lists and Nested or Split Data Frames}
\description{
Recursive lists in the form of R objects, JSON, and XML, for use in
teaching and examples. Examples include color palettes, Game of Thrones
characters, and GitHub users and repos. Data from the Gapminder package, as
a simple data frame and in nested and split forms.
}
\details{
Details
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/repurrrsive}
  \item Report bugs at \url{https://github.com/jennybc/repurrrsive/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@stat.ubc.ca}

}

