% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\docType{data}
\name{wesanderson}
\alias{wesanderson}
\title{Color palettes from Wes Anderson movies}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{https://cran.r-project.org/package=wesanderson}

\url{http://wesandersonpalettes.tumblr.com}
}
\usage{
wesanderson
}
\description{
A list of color palettes inspried by Wes Anderson movies, taken from the from
\code{\link{wesanderson}} package.
}
\keyword{datasets}

