% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\name{wesanderson_json}
\alias{wesanderson_json}
\title{Path to wesanderson JSON}
\usage{
wesanderson_json()
}
\value{
Local path to JSON file containing Wes Anderson color palettes
}
\description{
Path to wesanderson JSON
}
\examples{
wesanderson_json()
if (require("jsonlite")) {
jsonlite::fromJSON(wesanderson_json())
}
}

