% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\name{wesanderson_xml}
\alias{wesanderson_xml}
\title{Path to wesanderson XML}
\usage{
wesanderson_xml()
}
\value{
Local path to XML file containing Wes Anderson color palettes
}
\description{
Path to wesanderson XML
}
\examples{
wesanderson_xml()
if (require("xml2")) {
read_xml(wesanderson_xml())
}
}

