% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repurrrsive-package.R
\docType{package}
\name{repurrrsive-package}
\alias{repurrrsive}
\alias{repurrrsive-package}
\title{repurrrsive: Examples of Recursive Lists and Nested or Split Data Frames}
\description{
Recursive lists in the form of R objects, 'JSON', and 'XML', for use in teaching and examples. Examples include color palettes, Game of Thrones characters, 'GitHub' users and repositories, music collections, and entities from the Star Wars universe. Data from the 'gapminder' package is also included, as a simple data frame and in nested and split forms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jennybc.github.io/repurrrsive/}
  \item \url{https://github.com/jennybc/repurrrsive}
  \item Report bugs at \url{https://github.com/jennybc/repurrrsive/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@posit.co} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Other contributors:
\itemize{
  \item Charlotte Wickham \email{cwickham@gmail.com} [contributor]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
