% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repurrrsive-package.r
\docType{package}
\name{repurrrsive-package}
\alias{repurrrsive}
\alias{repurrrsive-package}
\title{repurrrsive: Recursive List Examples}
\description{
Recursive lists in the form of R objects, JSON, and XML, for use in
teaching and examples. Includes color palettes, GitHub user and repo data,
and SOMETHING ELSE.
}
\details{
Details
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jennybc/repurrrsive}
  \item Report bugs at \url{https://github.com/jennybc/repurrrsive/issues}
}

}
\author{
\strong{Maintainer}: Jennifer Bryan \email{jenny@stat.ubc.ca}

}

