% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\docType{data}
\name{wesanderson}
\alias{wesanderson}
\title{Color palettes from Wes Anderson movies}
\format{
A named list with 15 components, each containing a color palette from
a specific movie. Each palette consists of 4 or 5 hexadecimal color values.
}
\source{
\url{https://cran.r-project.org/package=wesanderson}

\url{http://wesandersonpalettes.tumblr.com}
}
\usage{
wesanderson
}
\description{
A list of color palettes inspired by Wes Anderson movies, taken from the from
\link[wesanderson:wesanderson]{wesanderson} package.
}
\examples{
wesanderson[4:5]
wesanderson[["Rushmore"]]
}
\seealso{
Other wesanderson data and functions: 
\code{\link{wesanderson_json}()}
}
\concept{wesanderson data and functions}
\keyword{datasets}
