% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\name{wesanderson_json}
\alias{wesanderson_json}
\alias{wesanderson_xml}
\title{Path to wesanderson JSON and XML}
\usage{
wesanderson_json()

wesanderson_xml()
}
\value{
Local path to JSON or XML file containing Wes Anderson color palettes
}
\description{
Path to wesanderson JSON and XML
}
\examples{
wesanderson_json()
if (require("jsonlite")) {
  jsonlite::fromJSON(wesanderson_json())
}
wesanderson_xml()
if (require("xml2")) {
  xml2::read_xml(wesanderson_xml())
}
}
\seealso{
Other wesanderson data and functions: \code{\link{wesanderson}}
}
